requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==2.3.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.2.0
pymongo==3.11.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
Werkzeug==2.3.7
wsproto==1.2.0